# specs/002-execution-intelligence-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Cipla EMEU Execution Intelligence API
  version: 0.1.0
  description: Read-only dashboard API plus grounded AI query endpoint for the protected MVP.
servers:
  - url: http://localhost:8000
    description: Local backend
paths:
...
  /api/workflow/summary:
    get:
      summary: Request and post-event workflow governance summary
      parameters:
        - $ref: "#/components/parameters/Country"
        - $ref: "#/components/parameters/Month"
        - $ref: "#/components/parameters/InterventionType"
        - $ref: "#/components/parameters/IncludeOutOfScope"
      responses:
        "200":
          description: Request approval, confirmation, and reporting status summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowSummaryResponse"
  /api/workflow/requests:
    get:
      summary: Workflow governance request-level rows
      parameters:
        - $ref: "#/components/parameters/Country"
        - $ref: "#/components/parameters/Month"
        - $ref: "#/components/parameters/InterventionType"
        - $ref: "#/components/parameters/IncludeOutOfScope"
        - name: workflowStatus
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated workflow governance rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRequestsResponse"
...
  /api/execution/summary:
    get:
      summary: Execution KPI summary
      parameters:
        - $ref: "#/components/parameters/Country"
        - $ref: "#/components/parameters/Month"
        - $ref: "#/components/parameters/IncludeOutOfScope"
        - $ref: "#/components/parameters/Therapy"
        - $ref: "#/components/parameters/EventType"
      responses:
        "200":
          description: Execution KPIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionSummaryResponse"
  /api/execution/events:
    get:
      summary: Event-level execution matrix
      parameters:
        - $ref: "#/components/parameters/Country"
        - $ref: "#/components/parameters/Month"
        - $ref: "#/components/parameters/IncludeOutOfScope"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
...
components:
  parameters:
    Country:
      name: country
      in: query
      schema:
        type: string
    Month:
      name: month
      in: query
      schema:
        type: string
        example: "2026-05"
...
    Therapy:
      name: therapy
      in: query
      schema:
        type: string
    EventType:
      name: eventType
      in: query
      schema:
        type: string
    InterventionType:
      name: interventionType
      in: query
      schema:
        type: string
    IncludeOutOfScope:
      name: includeOutOfScope
      in: query
      schema:
        type: boolean
        default: false
      description: Include audit-only rows outside Nepal/Sri Lanka Apr-May 2026 Phase 4 production scope.
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 25
        minimum: 1
        maximum: 100
  schemas:
...
    ResponseMeta:
      type: object
      required: [generatedAt, latestIngestionStatus, dataQualityFlags, limitations]
      properties:
        generatedAt:
          type: string
          format: date-time
        latestIngestionRunId:
          type: string
          nullable: true
        latestIngestionStatus:
          type: string
        dataQualityFlags:
          type: array
          items:
            type: string
        limitations:
          type: array
          items:
            type: string
        sourceDerivationNotes:
          type: array
          items:
            type: string
...
    ExecutionSummaryResponse:
      type: object
      required: [meta, plannedEvents, matchedEvents, executedEvents, actionDueEvents, matchCoverage]
      properties:
        meta:
          $ref: "#/components/schemas/ResponseMeta"
        plannedEvents:
          type: integer
        matchedEvents:
          type: integer
        weakOrUnmatchedEvents:
          type: integer
        executedEvents:
          type: integer
        actionDueEvents:
          type: integer
        plannedEventsWithExecutedEvidence:
          type: integer
        plannedEventsWithActionDueEvidence:
          type: integer
        executedSnapshotCount:
          type: integer
        actionDueSnapshotCount:
          type: integer
        plannedHcps:
          type: integer
        engagedHcps:
          type: integer
        matchedEngagedHcps:
          type: integer
        rawEngagedHcps:
          type: integer
        hcpExecutionRate:
          type: number
        eventExecutionRate:
          type: number
        matchCoverage:
          type: number
        snapshotSourceCounts:
          type: object
          additionalProperties:
            type: integer
        primaryScope:
          type: boolean
        scopeStatuses:
          type: array
          items:
            type: string
        scopeReasons:
          type: array
          items:
            type: string
    EventListResponse:
      type: object
      required: [meta, page, pageSize, total, rows]
      properties:
        meta:
          $ref: "#/components/schemas/ResponseMeta"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        rows:
          type: array
          items:
            type: object
...
    WorkflowSummaryResponse:
      type: object
      required: [meta, requestApprovalCounts, requestConfirmationCounts, postApprovalCounts, postConfirmationCounts, ownerStageCounts]
      properties:
        meta:
          $ref: "#/components/schemas/ResponseMeta"
        requestApprovalCounts:
          type: object
          additionalProperties:
            type: integer
        requestConfirmationCounts:
          type: object
          additionalProperties:
            type: integer
        postApprovalCounts:
          type: object
          additionalProperties:
            type: integer
        postConfirmationCounts:
          type: object
          additionalProperties:
            type: integer
        ownerStageCounts:
          type: object
          additionalProperties:
            type: integer
        pendingRequestCount:
          type: integer
        pendingReportCount:
          type: integer
        reportsSentForCorrection:
          type: integer
        reportsApproved:
          type: integer
        expenseSubmittedCoverage:
          type: number
        expenseConfirmedCoverage:
          type: number
        primaryScope:
          type: boolean
        scopeStatuses:
          type: array
          items:
            type: string
        scopeReasons:
          type: array
          items:
            type: string
    WorkflowRequestsResponse:
      type: object
      required: [meta, page, pageSize, total, rows]
      properties:
        meta:
          $ref: "#/components/schemas/ResponseMeta"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        rows:
          type: array
          items:
            type: object
            properties:
              reqId:
                type: string
              country:
                type: string
              month:
                type: string
              repName:
                type: string
                nullable: true
              interventionType:
                type: string
                nullable: true
              requestApprovalStatus:
                type: string
              requestConfirmationStatus:
                type: string
              postApprovalStatus:
                type: string
              postConfirmationStatus:
                type: string
              currentOwnerStage:
                type: string
                nullable: true
              isPrimaryPhase4Scope:
                type: boolean
              scopeStatus:
                type: string
                nullable: true
              scopeReason:
                type: string
                nullable: true
...